requirements.txt
langchain==0.1.8
langchain-openai==0.0.6
fastapi==0.111.0
numpy~=1.26.0
openai==1.30.4
google-generativeai
pandas~=2.1.1
plotly~=5.18.0
pgvector==0.2.5
psycopg2-binary==2.9.9
python-dotenv~=1.0.0
pypdf2~=3.0.1
langchain_google_genai
scipy~=1.12.0
scikit-learn~=1.4.1.post1
SQLAlchemy~=1.4.49
langchain
tiktoken==0.6.0
requests-toolbelt==1.0.0
google-generativeai
uvicorn==0.29.0
python-docx==1.1.0
python-multipart==0.0.9
sentence-transformers==2.2.2
transformers==4.39.0
pytransformers==0.1.0
bson
selenium
bs4